requirements.txt
python-dotenv==1.0.0
openai==0.27.2
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.10.1
matplotlib==3.7.1
plotly==5.13.1
